chore: Restructures the import module
